.bash_completion.d/adb.bash: fix completion for white spaces in filename .bashrc: add per pc customization bin/adb-browse-with-pc: fix weixin browse with pc bin/adb-get-a-note: get a note from t1 bin/adb-get-last-pic: allow remove the pic after get it, and redirect output so to be usable bin/adb-search-contact: search for contact, to work around it always lost after click in bin/adb-settings: add more settings bin/adb-start-vpn: start the vpn without waiting bin/adb-weixin-choose-chatter: optimization bin/adb-weixin-choose-chatter: make 黄 波 to 黄.*波 bin/after-co-ln-s.sh: do not replace it blindly bin/do-remote-system-config: should not do system config sometimes, only copy the id_rsas bin/overide/adb: add a start-activity sub command bin/smb2ftp: smbcp first, then lftp bin/vm: reverse mv

+++ bin/after-co-ln-s.sh
@@ -204,7 +204,12 @@ if ask-if-not-bhj "Do you want to switch the ctrl/alt, esc/caps_lock keys?"; the
         relative-link -f ~/etc/hardware-mach/$mach/.Xmodmap ~/.Xmodmap
     fi
 fi
-sudo ln -sf ~/etc/rc.local /etc || true # no sudo on win32
+if test -L /etc/rc.local || yes-or-no-p -N "Replace /etc/rc.local with system-config's version?"; then
+    if ! test -L /etc/rc.local; then
+        cp /etc/rc.local /etc/rc.local.bak
+    fi
+    sudo ln -sf ~/etc/rc.local /etc || true # no sudo on win32
+fi
 mkdir -p ~/bin/$(uname|perl -npe 's/_.*//')/ext/`uname -m`/
 if test -L ~/.git; then rm -f ~/.git; fi
 if test $(uname) = Linux; then
